refactor(mem): static cleanup — remove dead code, silence int64 format warnings (#213)
Output-preserving static cleanup. Verified byte-identical: full `bwa-mem3 mem`
SAM output on 1M single-end and 500k paired-end reads (md5 unchanged vs
origin/main modulo the @PG tag), plus the sw-kernel-bench byte-identity golden
gate (extend + rescue, 60k pairs each, GOLDEN CHECK PASS on the neon tier) for
the banded-SW file.

Dead code removed:
- bandedSWA.cpp: unread minLen1/minLen2 locals in the 8-bit and 16-bit banded-SW
  batch wrappers. The 8-bit uint8_t pair also tripped -Woverflow (MAX_SEQ_LEN8+1
  does not fit uint8_t).
- bwamem.cpp: the never-called PDQSORT_INIT(mem_flt) wrapper, and the fully
  unused KSORT_INIT/PDQSORT_INIT(mem_intv) pair plus their intv_lt comparator.
  Their doc comment described mem_collect_intv, a function that no longer exists.
  Only the live mem_flt ks_introsort and the mem_intv1 SMEM sort remain.
- FMI_search.{cpp,h}: the dead getSMEMs method, superseded by the
  getSMEMs*OneThread family; it had no callers.

Format warnings (-Wformat) fixed with the repo's existing (long long)+%lld idiom
for int64_t diagnostics: bandedSWA getTicks(), plus stderr/verbose prints in
bwamem.cpp, fastmap.cpp, and FMI_search.cpp. Diagnostic paths only; no effect on
alignment output. Drops the build from 17 -Wformat warnings to 0.

+++ src/fastmap.cpp
@@ -376,8 +376,8 @@ ktp_data_t *kt_pipeline(void *shared, int step, void *data, mem_opt_t *opt, work
             ret->prof.n_bp = sz;
         }
 
-        fprintf(stderr, "[0000] read_chunk: %ld, work_chunk_size: %ld, nseq: %d\n",
-                aux->task_size, sz, ret->n_seqs);
+        fprintf(stderr, "[0000] read_chunk: %lld, work_chunk_size: %lld, nseq: %d\n",
+                (long long)aux->task_size, (long long)sz, ret->n_seqs);
 
         if (ret->seqs == 0) {
             free(ret);
